+++ head/lib/Head.pm
@@ -15,7 +15,7 @@ use Head::Command::checkclients;
 
 use Sys::Hostname;
 
-our $VERSION = '2.74';
+our $VERSION = '2.75';
 
 # This method will run once at server start
 sub startup {